+++ examples/index.tsx
@@ -1,0 +1,6 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import App from './app/App';
+
+// @ts-ignore
+ReactDOM.render( <App version={SWIPEABLE_VERSION} />, document.getElementById('app') );