+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+# v6.0.0
+**New Features:**
+- include passive event listener option, by default, to internal uses of `addEventListener`
+  - solves issue with chrome and lighthouse - [#167](https://github.com/FormidableLabs/react-swipeable/issues/167)
+  - set `passive` to `false` only when `preventDefaultTouchmoveEvent` is `true`.
+  - more details in [readme#passive-listener-issue](https://github.com/FormidableLabs/react-swipeable#passive-listener)
+- add new `onTap` event handler prop which executes its callback after a tap
+  - Thank you [@upatel32](https://github.com/upatel32)!
+- add new `vxvy` event data property
+  - `[ deltaX/time, deltaY/time]` - velocity per axis
+  - Thank you [@upatel32](https://github.com/upatel32)!
+
+**Breaking Changes:**
+- **remove** `<Swipeable>` component
+  - see below for an example of how to make your own
+  - [Swipeable component examples](https://github.com/FormidableLabs/react-swipeable/blob/main/migration.md#swipeable-component-examples)
+- **event data update** correctly calculate `deltaX` and `deltaY`
+  - from `initial - current` **to** `current - initial`
+  - fixes issue [#157](https://github.com/FormidableLabs/react-swipeable/issues/157)
+  - Thank you [@upatel32](https://github.com/upatel32)!
+- **drop support for ie11**
+  - using `addEventListener` options object needs to be polyfilled, [browser support](https://github.com/FormidableLabs/react-swipeable#browser-support)
+- **requires** react >= 16.8.3, additionally supports new react v17
+
+**Bug fixes:**
+- Swipes can now start at edges (x or y === 0)
+  - fixes [#182](https://github.com/FormidableLabs/react-swipeable/issues/182)
+  - Thank you [@upatel32](https://github.com/upatel32)!
+
+**Infrastructure:**
+- **typescript** Converted entire code base, tests, and examples to typescript
+  - **changed type** `EventData` -> `SwipeEventData` - The event data provided for all swipe event callbacks
+  - **removed type** `SwipeableOptions` - use `SwipeableProps` now
+  - **removed types** associated with `<Swipeable>` component
+  - **new type** `TapCallback` - callback for the new `onTap` prop handler
+  - **new type** `SwipeDirections` - `"Left" | "Right" | "Up" | "Down"`
+- Converted tests to `@testing-library/react`, [react testing library](https://github.com/testing-library/react-testing-library)
+- Build bundles with `microbundle`. [microbundle](https://github.com/developit/microbundle)
+  - export new "modern" build - via package.json `esmodule` property
+    - [microbundle modern mode](https://github.com/developit/microbundle#-modern-mode-)
+
+**Maintenance:**
+- Upgraded all dev dependencies, `jest`, `babel`, `webpack`, `eslint`, `prettier`
+
 # 5.5.0
 * Add `first` property to `eventData` that is `true` for first swipe event [issue #160](https://github.com/formidablelabs/react-swipeable/issues/160) and [PR #162](https://github.com/formidablelabs/react-swipeable/pull/162)
   * Thank you [@samanpwbb](https://github.com/samanpwbb)!